chore: remove nullability and ability to use Integer for PlanningPinToIndex
